modules/auth/requirements/common.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.6
django-allauth==0.42.0
django-filter==2.2.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
idna==2.9
importlib-metadata==1.6.0
Markdown==3.2.2
oauthlib==3.1.0
psycopg2==2.8.5
PyJWT==1.7.1
python-decouple==3.3
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
zipp==3.1.0
